Remove `CommandHandler` interface in favor of the abstract class
