+++ server/src/main/java/org/spine3/server/command/CommandHandler.java
@@ -20,11 +20,27 @@
 
 package org.spine3.server.command;
 
+import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Any;
+import com.google.protobuf.Message;
+import com.google.protobuf.StringValue;
+import org.spine3.base.CommandClass;
+import org.spine3.base.CommandContext;
+import org.spine3.base.CommandEnvelope;
+import org.spine3.base.Event;
+import org.spine3.protobuf.AnyPacker;
 import org.spine3.server.event.EventBus;
+import org.spine3.server.reflect.CommandHandlerMethod;
+
+import javax.annotation.Nullable;
+import java.util.List;
+import java.util.Set;
+
+import static org.spine3.protobuf.Values.newStringValue;
 
 /**
- * A base interface for non-aggregate objects that handle commands.
+ * The abstract base for non-aggregate classes that expose command handling methods
+ * and post their results to {@link EventBus}.
  *
  * <p>A command handler is responsible for:
  * <ol>
@@ -42,19 +58,122 @@ import org.spine3.server.event.EventBus;
  * some rare cases, using {@linkplain org.spine3.server.aggregate.Aggregate aggregates}
  * is a preferred way of handling commands.
  *
+ * <p>This class implements {@code CommandDispatcher} dispatching messages
+ * to methods declared in the derived classes.
+ *
  * @author Alexander Yevsyukov
+ * @see org.spine3.server.aggregate.Aggregate Aggregate
+ * @see CommandDispatcher
  */
-interface CommandHandler extends CommandDispatcher {
+public abstract class CommandHandler implements CommandDispatcher {
 
     /**
-     * Obtains identifier for being placed in to {@link org.spine3.base.EventContext EventContext}
-     * of the produced events.
+     * The {@code EventBut} to which the handler posts events it produces.
      */
-    Any getProducerId();
+    private final EventBus eventBus;
 
     /**
-     * Obtains the reference to {@code EventBus} to which the handler posts the
-     * produced events.
+     * Fully qualified name of the class wrapped into {@code Any}.
      */
-    EventBus getEventBus();
+    private final Any producerId;
+
+    /**
+     * Cached set of the command classes this handler can handle.
+     */
+    @Nullable
+    private Set<CommandClass> commandClasses;
+
+    /**
+     * Creates a new instance of the command handler.
+     *
+     * @param eventBus the {@code EventBus} to post events generated by this handler
+     */
+    protected CommandHandler(EventBus eventBus) {
+        this.eventBus = eventBus;
+        final StringValue className = newStringValue(getClass().getName());
+        this.producerId = AnyPacker.pack(className);
+    }
+
+    /**
+     * Dispatches the command to the handler method and
+     * posts resulting events to the {@link EventBus}.
+     *
+     * @param envelope the command to dispatch
+     * @throws IllegalStateException if an exception occurred during command dispatching
+     *                               with this exception as the cause
+     */
+    @Override
+    public void dispatch(CommandEnvelope envelope) {
+        handle(envelope.getMessage(), envelope.getCommandContext());
+    }
+
+    @SuppressWarnings("ReturnOfCollectionOrArrayField") // OK as we return immutable impl.
+    @Override
+    public Set<CommandClass> getMessageClasses() {
+        if (commandClasses == null) {
+            commandClasses = ImmutableSet.copyOf(
+                    CommandHandlerMethod.getCommandClasses(getClass()));
+        }
+        return commandClasses;
+    }
+
+    /**
+     * Dispatches the command to the handler method and
+     * posts resulting events to the {@link EventBus}.
+     *
+     * @param commandMessage the command message
+     * @param context        the command context
+     * @throws IllegalStateException if an exception occurred during command dispatching
+     *                               with this exception as the cause
+     */
+    private void handle(Message commandMessage, CommandContext context) {
+        final List<? extends Message> eventMessages =
+                CommandHandlerMethod.invokeHandler(this, commandMessage, context);
+        final List<Event> events = toEvents(eventMessages, context);
+        postEvents(events);
+    }
+
+    private List<Event> toEvents(List<? extends Message> eventMessages,
+                                 CommandContext commandContext) {
+        return CommandHandlerMethod.toEvents(producerId,
+                                             null,
+                                             eventMessages,
+                                             commandContext);
+    }
+
+    /** Posts passed events to {@link EventBus}. */
+    private void postEvents(Iterable<Event> events) {
+        for (Event event : events) {
+            eventBus.post(event);
+        }
+    }
+
+    /**
+     * Indicates whether some another command handler is “equal to” this one.
+     *
+     * <p>Two command handlers are equal if they handle the same set of commands.
+     *
+     * @return if the passed {@code CommandHandler} handles the same
+     * set of command classes.
+     * @see #getMessageClasses()
+     */
+    @Override
+    public boolean equals(Object otherObj) {
+        if (this == otherObj) {
+            return true;
+        }
+        if (otherObj == null ||
+            getClass() != otherObj.getClass()) {
+            return false;
+        }
+        final CommandHandler otherHandler = (CommandHandler) otherObj;
+        final boolean equals = getMessageClasses().equals(otherHandler.getMessageClasses());
+        return equals;
+    }
+
+    @Override
+    public int hashCode() {
+        final int result = getMessageClasses().hashCode();
+        return result;
+    }
 }
